added automatic preview generation on save
