--- src/utils/MascotState.ts
@@ -1,4 +1,0 @@
-import { useState, createContext } from "react";
-import { EEmotion } from "../components/logic/EEmotion";
-import IMascot from "../components/logic/IMascot";
-
